Moved events to events class
